Support large network topologies with long integer capacities
Refactor capacity-related methods and fields in `UVNode`, `LNode`, and related classes to use `long` instead of `int`. Update calculations to handle large values without overflow. Add regression tests to validate topology imports and ensure proper handling of capacities exceeding `Integer.MAX_VALUE`.

+++ src/UltraViolet.java
@@ -1429,7 +1429,7 @@ public class UltraViolet {
         return width;
     }
 
-    private String formatAmountWithUnit(int amount, int width) {
+    private String formatAmountWithUnit(long amount, int width) {
         return padLeft(formatAmountNumber(amount), width) + " sats";
     }
 
@@ -1471,7 +1471,7 @@ public class UltraViolet {
     }
 
     private double safeOverallOutboundFraction(UVNode node) {
-        int capacity = node.getNodeCapacity();
+        long capacity = node.getNodeCapacity();
         if (capacity == 0) {
             return 0.0;
         }
